+++ src/spawn_handle.rs
@@ -2,145 +2,165 @@
 //
 use
 {
-	crate        :: { import::*, JoinHandle, remote_handle::remote_handle              } ,
-	futures_util :: { task::{ SpawnExt, SpawnError }, future::{ FutureExt, abortable } } ,
-	std          :: { future::Future, sync::{ Arc, atomic::AtomicBool }, rc::Rc        } ,
+	futures_util :: { future::{ FutureExt, abortable }, task::SpawnExt                    } ,
+	futures_task :: { SpawnError, FutureObj                                               } ,
+	crate        :: { import::*, JoinHandle, remote_handle::remote_handle                 } ,
+	std          :: { pin::Pin, future::Future, sync::{ Arc, atomic::AtomicBool }, rc::Rc } ,
 };
 
 
 /// Let's you spawn and get a [JoinHandle] to await the output of a future.
 ///
-/// This trait is not object safe, see [SpawnHandleOs](crate::SpawnHandleOs) for a best effort object safe one.
+/// This trait works much like the [`Spawn`](futures_task::Spawn) trait from the futures library.
+/// It takes a [`FutureObj`] so we can hopefully make it `no_std` compatible when needed. This
+/// also allows it to be object safe.  For convenience, there is [`SpawnHandleExt`] which allows you
+/// to spawn a generic future directly without having to manually make the [`FutureObj`].
 ///
-/// ## Performance
+/// [`SpawnHandleExt`] is automatically implemented but must be in scope, so this works:
 ///
-/// For [tokio] and [async-std](async_std_crate) this is generally faster than [SpawnExt::spawn], since
-/// it's better aligned with the underlying API and doesn't require extra boxing.
+/// ```rust
+/// use async_executors::{ SpawnHandle, SpawnHandleExt };
+///
+/// async fn need_exec( exec: impl SpawnHandle<()> )
+/// {
+///    let join_handle = exec.spawn_handle( async {} ).expect( "spawn" );
+///
+///    join_handle.await;
+/// }
+/// ```
+///
+/// and so does this:
+///
+/// ```rust
+/// use async_executors::{ SpawnHandle, SpawnHandleExt };
+///
+/// async fn need_exec( exec: Box< dyn SpawnHandle<()> > )
+/// {
+///    let join_handle = exec.spawn_handle( async {} ).expect( "spawn" );
+///
+///    join_handle.await;
+/// }
+/// ```
+///
+/// One inconvenience of it having to be object safe is that the trait needs to be generic over the
+/// output parameter. This can be annoying if you need an executor that can spawn futures with different
+/// output parameters. Normally you should always be able to know which ones you need. If not
+/// you will have to make the type that stores the executor generic over the output type as well.
+///
+/// So to enable several output types you can use the
+/// [following workaround](https://github.com/najamelan/async_executors/tree/master/examples/spawn_handle_multi.rs).
 //
 #[ cfg_attr( feature = "docs", doc(cfg( feature = "spawn_handle" )) ) ]
 //
-pub trait SpawnHandle
+pub trait SpawnHandle<Out: 'static + Send>
+{
+	/// Spawn a future and return a [`JoinHandle`] that can be awaited for the output of the future.
+	//
+	fn spawn_handle_obj( &self, future: FutureObj<'static, Out> ) -> Result<JoinHandle<Out>, SpawnError>;
+}
+
+/// Convenience trait for passing in a generic future to [`SpawnHandle`]. Much akin to `Spawn` and `SpawnExt` in the
+/// futures library.
+//
+#[ cfg_attr( feature = "docs", doc(cfg( feature = "spawn_handle" )) ) ]
+//
+pub trait SpawnHandleExt<Out: 'static + Send> : SpawnHandle<Out>
 {
 	/// Spawn a future and return a [JoinHandle] that can be awaited for the output of the future.
 	//
-	fn spawn_handle<Fut, Out>( &self, future: Fut ) -> Result<JoinHandle<Out>, SpawnError>
-
-		where Fut: Future<Output = Out> + 'static + Send,
-		      Out: 'static + Send
-	;
+	fn spawn_handle( &self, future: impl Future<Output = Out> + Send + 'static ) -> Result<JoinHandle<Out>, SpawnError>;
 }
 
 
-impl<T: SpawnHandle> SpawnHandle for Box<T>
+impl<T, Out> SpawnHandleExt<Out> for T
+
+	where T  : SpawnHandle<Out> + ?Sized ,
+	      Out: 'static + Send            ,
 {
-	fn spawn_handle<Fut, Out>( &self, future: Fut ) -> Result<JoinHandle<Out>, SpawnError>
-
-		where Fut: Future<Output = Out> + 'static + Send,
-		      Out: 'static + Send
-
+	fn spawn_handle( &self, future: impl Future<Output = Out> + Send + 'static ) -> Result<JoinHandle<Out>, SpawnError>
 	{
-		(**self).spawn_handle( future )
+		self.spawn_handle_obj( FutureObj::new(future.boxed()) )
 	}
 }
 
 
-impl<T: SpawnHandle> SpawnHandle for Arc<T>
+impl<T: ?Sized, Out> SpawnHandle<Out> for Box<T> where T: SpawnHandle<Out>, Out: 'static + Send
 {
-	fn spawn_handle<Fut, Out>( &self, future: Fut ) -> Result<JoinHandle<Out>, SpawnError>
-
-		where Fut: Future<Output = Out> + 'static + Send,
-		      Out: 'static + Send
-
+	fn spawn_handle_obj( &self, future: FutureObj<'static, Out> ) -> Result<JoinHandle<Out>, SpawnError>
 	{
-		(**self).spawn_handle( future )
+		(**self).spawn_handle_obj( future )
 	}
 }
 
 
-impl<T: SpawnHandle> SpawnHandle for Rc<T>
+impl<T: ?Sized, Out> SpawnHandle<Out> for Arc<T> where T: SpawnHandle<Out>, Out: 'static + Send
 {
-	fn spawn_handle<Fut, Out>( &self, future: Fut ) -> Result<JoinHandle<Out>, SpawnError>
-
-		where Fut: Future<Output = Out> + 'static + Send,
-		      Out: 'static + Send
-
+	fn spawn_handle_obj( &self, future: FutureObj<'static, Out> ) -> Result<JoinHandle<Out>, SpawnError>
 	{
-		(**self).spawn_handle( future )
+		(**self).spawn_handle_obj( future )
 	}
 }
 
 
-impl<T: SpawnHandle> SpawnHandle for &T
+impl<T: ?Sized, Out> SpawnHandle<Out> for Rc<T> where T: SpawnHandle<Out>, Out: 'static + Send
 {
-	fn spawn_handle<Fut, Out>( &self, future: Fut ) -> Result<JoinHandle<Out>, SpawnError>
-
-		where Fut: Future<Output = Out> + 'static + Send,
-		      Out: 'static + Send
-
+	fn spawn_handle_obj( &self, future: FutureObj<'static, Out> ) -> Result<JoinHandle<Out>, SpawnError>
 	{
-		(**self).spawn_handle( future )
+		(**self).spawn_handle_obj( future )
 	}
 }
 
 
-impl<T: SpawnHandle> SpawnHandle for &mut T
+impl<T, Out> SpawnHandle<Out> for &T where T: SpawnHandle<Out>, Out: 'static + Send
 {
-	fn spawn_handle<Fut, Out>( &self, future: Fut ) -> Result<JoinHandle<Out>, SpawnError>
-
-		where Fut: Future<Output = Out> + 'static + Send,
-		      Out: 'static + Send
-
+	fn spawn_handle_obj( &self, future: FutureObj<'static, Out> ) -> Result<JoinHandle<Out>, SpawnError>
 	{
-		(**self).spawn_handle( future )
+		(**self).spawn_handle_obj( future )
+	}
+}
+
+
+impl<T, Out> SpawnHandle<Out> for &mut T where T: SpawnHandle<Out>, Out: 'static + Send
+{
+	fn spawn_handle_obj( &self, future: FutureObj<'static, Out> ) -> Result<JoinHandle<Out>, SpawnError>
+	{
+		(**self).spawn_handle_obj( future )
 	}
 }
 
 
 #[ cfg( feature = "tracing" ) ]
 //
-impl<T: SpawnHandle> SpawnHandle for Instrumented<T>
+impl<T, Out> SpawnHandle<Out> for Instrumented<T> where T: SpawnHandle<Out>, Out: 'static + Send
 {
-	fn spawn_handle<Fut, Out>( &self, future: Fut ) -> Result<JoinHandle<Out>, SpawnError>
-
-		where Fut: Future<Output = Out> + 'static + Send,
-		      Out: 'static + Send
-
+	fn spawn_handle_obj( &self, future: FutureObj<'static, Out> ) -> Result<JoinHandle<Out>, SpawnError>
 	{
 		let fut = future.instrument( self.span().clone() );
 
-		self.inner().spawn_handle( fut )
+		self.inner().spawn_handle_obj( FutureObj::new(fut.boxed()) )
 	}
 }
 
 
 #[ cfg( feature = "tracing" ) ]
 //
-impl<T: SpawnHandle> SpawnHandle for WithDispatch<T>
+impl<T, Out> SpawnHandle<Out> for WithDispatch<T> where T: SpawnHandle<Out>, Out: 'static + Send
 {
-	fn spawn_handle<Fut, Out>( &self, future: Fut ) -> Result<JoinHandle<Out>, SpawnError>
-
-		where Fut: Future<Output = Out> + 'static + Send,
-		      Out: 'static + Send
-
+	fn spawn_handle_obj( &self, future: FutureObj<'static, Out> ) -> Result<JoinHandle<Out>, SpawnError>
 	{
 		let fut = self.with_dispatch( future );
 
-		self.inner().spawn_handle( fut )
+		self.inner().spawn_handle_obj( FutureObj::new(fut.boxed()) )
 	}
 }
-
 
 
 
 #[ cfg( feature = "async_std" ) ]
 //
-impl SpawnHandle for crate::async_std::AsyncStd
+impl<Out: 'static + Send> SpawnHandle<Out> for crate::async_std::AsyncStd
 {
-	fn spawn_handle<Fut, Out>( &self, future: Fut ) -> Result<JoinHandle<Out>, SpawnError>
-
-		where Fut: Future<Output = Out> + 'static + Send,
-		      Out: 'static + Send
-
+	fn spawn_handle_obj( &self, future: FutureObj<'static, Out> ) -> Result<JoinHandle<Out>, SpawnError>
 	{
 		let (fut, a_handle) = abortable( future );
 
@@ -154,15 +174,12 @@ impl SpawnHandle for crate::async_std::AsyncStd
 }
 
 
+
 #[ cfg(any( feature = "tokio_tp", feature = "tokio_ct" )) ]
 //
-impl SpawnHandle for crate::TokioHandle
+impl<Out: 'static + Send> SpawnHandle<Out> for crate::TokioHandle
 {
-	fn spawn_handle<Fut, Out>( &self, future: Fut ) -> Result<JoinHandle<Out>, SpawnError>
-
-		where Fut: Future<Output = Out> + 'static + Send,
-		      Out: 'static + Send
-
+	fn spawn_handle_obj( &self, future: FutureObj<'static, Out> ) -> Result<JoinHandle<Out>, SpawnError>
 	{
 		let (fut, a_handle) = abortable( future );
 
@@ -176,15 +193,12 @@ impl SpawnHandle for crate::TokioHandle
 }
 
 
+
 #[ cfg( feature = "tokio_tp" ) ]
 //
-impl SpawnHandle for crate::TokioTp
+impl<Out: 'static + Send> SpawnHandle<Out> for crate::TokioTp
 {
-	fn spawn_handle<Fut, Out>( &self, future: Fut ) -> Result<JoinHandle<Out>, SpawnError>
-
-		where Fut: Future<Output = Out> + 'static + Send,
-		      Out: 'static + Send
-
+	fn spawn_handle_obj( &self, future: FutureObj<'static, Out> ) -> Result<JoinHandle<Out>, SpawnError>
 	{
 		let (fut, a_handle) = abortable( future );
 
@@ -199,15 +213,11 @@ impl SpawnHandle for crate::TokioTp
 
 
 
-#[ cfg(any( feature = "tokio_ct" )) ]
+#[ cfg( feature = "tokio_ct" ) ]
 //
-impl SpawnHandle for crate::TokioCt
+impl<Out: 'static + Send> SpawnHandle<Out> for crate::TokioCt
 {
-	fn spawn_handle<Fut, Out>( &self, future: Fut ) -> Result<JoinHandle<Out>, SpawnError>
-
-		where Fut: Future<Output = Out> + 'static + Send,
-		      Out: 'static + Send
-
+	fn spawn_handle_obj( &self, future: FutureObj<'static, Out> ) -> Result<JoinHandle<Out>, SpawnError>
 	{
 		let (fut, a_handle) = abortable( future );
 
@@ -224,13 +234,9 @@ impl SpawnHandle for crate::TokioCt
 
 #[ cfg( feature = "bindgen" ) ]
 //
-impl SpawnHandle for crate::Bindgen
+impl<Out: 'static + Send> SpawnHandle<Out> for crate::Bindgen
 {
-	fn spawn_handle<Fut, Out>( &self, future: Fut ) -> Result<JoinHandle<Out>, SpawnError>
-
-		where Fut: Future<Output = Out> + 'static + Send,
-		      Out: 'static + Send
-
+	fn spawn_handle_obj( &self, future: FutureObj<'static, Out> ) -> Result<JoinHandle<Out>, SpawnError>
 	{
 		let (fut, handle) = remote_handle( future );
 		wasm_bindgen_futures::spawn_local(fut);
@@ -243,13 +249,9 @@ impl SpawnHandle for crate::Bindgen
 
 #[ cfg( feature = "localpool" ) ]
 //
-impl SpawnHandle for futures_executor::LocalSpawner
+impl<Out: 'static + Send> SpawnHandle<Out> for futures_executor::LocalSpawner
 {
-	fn spawn_handle<Fut, Out>( &self, future: Fut ) -> Result<JoinHandle<Out>, SpawnError>
-
-		where Fut: Future<Output = Out> + 'static + Send,
-		      Out: 'static + Send
-
+	fn spawn_handle_obj( &self, future: FutureObj<'static, Out> ) -> Result<JoinHandle<Out>, SpawnError>
 	{
 		let (fut, handle) = remote_handle( future );
 
@@ -263,13 +265,9 @@ impl SpawnHandle for futures_executor::LocalSpawner
 
 #[ cfg( feature = "threadpool" ) ]
 //
-impl SpawnHandle for futures_executor::ThreadPool
+impl<Out: 'static + Send> SpawnHandle<Out> for futures_executor::ThreadPool
 {
-	fn spawn_handle<Fut, Out>( &self, future: Fut ) -> Result<JoinHandle<Out>, SpawnError>
-
-		where Fut: Future<Output = Out> + 'static + Send,
-		      Out: 'static + Send
-
+	fn spawn_handle_obj( &self, future: FutureObj<'static, Out> ) -> Result<JoinHandle<Out>, SpawnError>
 	{
 		let (fut, handle) = remote_handle( future );