+++ tests/tokio_tp.rs
@@ -9,7 +9,7 @@
 // ✔ pass a Arc<TokioTp> to a function that takes exec: `impl Spawn`
 // ✔ pass a     TokioTp  to a function that takes exec: `impl SpawnHandle`
 // ✔ pass a Arc<TokioTp> to a function that takes exec: `impl SpawnHandle`
-// ✔ pass a    &TokioTp  to a function that takes exec: `&dyn SpawnHandleOs`
+// ✔ pass a    &TokioTp  to a function that takes exec: `&dyn SpawnHandle`
 // ✔ pass a builder with some config set.
 //
 mod common;
@@ -145,7 +145,7 @@ fn test_spawn_handle_arc()
 }
 
 
-// pass a AsyncStd to a function that takes exec: `&dyn SpawnHandleOs`
+// pass a AsyncStd to a function that takes exec: `&dyn SpawnHandle`
 //
 #[ cfg( feature = "spawn_handle" ) ]
 //